fix: suppress package manager install message
`yo-environment` v4 prints a console message if a generator doesn't
change the `package.json` file for a project [^1]. e.g.,

```console
$ yo app/index.js --force
...
No change to package.json was detected. No package manager install will be executed.
```

It's superfluous for a generator that only deals with a README. In other
words, I found the new behavior annoying and wanted to minimize the logs
printed 😅.

So, this change disables the message by telling Yeoman that the
generator has a custom install task, and then not providing one. I wish
I could reference docs describing the behaviour, but I mainly got the
intended behaviour through trial & error, as well as sifting through the
code in `yeoman-generator` [^1][^2].

[^1]: https://github.com/yeoman/environment/blob/9ddabe49be9bcb7e1723bd9db379c5e1c7ae7885/src/package-manager.ts#L71
[^2]: https://github.com/yeoman/environment/blob/9ddabe49be9bcb7e1723bd9db379c5e1c7ae7885/src/environment-base.ts#L796